input improvements when xlive shell/cartographer captures input, accommodating all systems
- imgui handler code improvements

+++ src/Blam/saved_games/cartographer_player_profile/cartographer_player_profile_weapon_offsets.h
@@ -2,7 +2,7 @@
 
 /* enums */
 
-enum e_weapon_offset_weapon : int8
+enum e_weapon_offset_weapon : int32
 {
 	_weapon_offset_weapon_battle_rifle,
 	_weapon_offset_weapon_beam_rifle,
